Still working on Panda push environments. Reran DQN and IV-DQN with update requirements.txt
For the panda push environment result files with results in them, they use CEM and not PF. I made a mistake in the naming of the files in the save_data_CEM (for both using a model of the env and the env). For the MC runs using the env as a model for MPC simulations, I didn't rerun them yet since even if the CEM results replace the PF ones I just want to get an idea if the results are good or not.